backend/requirements.txt
# Church Games Backend Requirements

# Core
fastapi>=0.115.0  # Updated to support anyio 4.x (required by google-genai)
uvicorn[standard]>=0.24.0  # ASGI server for FastAPI
python-multipart==0.0.6  # Required for file uploads
python-dotenv==1.0.0

# Database - Async
aiomysql==0.2.0
cryptography==41.0.7  # Required for MySQL sha256_password/caching_sha2_password authentication

# Database - Initialization (sync)
mysql-connector-python==8.2.0

# File I/O - Async
aiofiles==23.2.1
aiohttp==3.9.1  # For downloading preview images

# AI Providers
openai==1.10.0
anthropic==0.24.0
google-generativeai==0.3.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2

# Production ASGI Server
uvicorn[standard]>=0.24.0  # ASGI server for FastAPI
gunicorn==21.2.0  # Optional: use with uvicorn workers if needed

# Rate Limiting
slowapi==0.1.9

